\name{ispiranteRanney-package}
\alias{ispiranteRanney-package}
\alias{ispiranteRanney}
\docType{package}
\title{Functions to analyze Ispirante data}
\description{Explore help-desk ticket data for ispirante.}
\details{
\tabular{ll}{Package: \tab ispiranteRanney\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-01-07\cr
Title: \tab Functions to analyze Ispirante data\cr
Author: \tab Steven Ranney\cr
Maintainer: \tab Steven H. Ranney <steven.ranney@gmail.com>\cr
Depends: \tab R (>= 3.1.2), ggplot2, scales, forecast, zoo, timeDate\cr
License: \tab GPL-3\cr}
}
\author{Steven Ranney}

\keyword{ package }


