\name{PlotReopensByDate}
\alias{PlotReopensByDate}
\title{Create a plot of the number of reopens by date}
\description{This function plots the number of reopens and produces a box 
plot of those values as a function of \code{date} level. }
\usage{PlotReopensByDate(dF, beginDate = min(dF$assignedDate), endDate = max(dF$assignedDate), 
    ...)}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by date level.}
  \item{beginDate}{A character that includes the beginning
date of interest.  Value \emph{must} 
be in the form \code{"yyyy-mm-dd"}, 
e.g., "2014-01-01". The function will 
default to the earliest date in the data frame.}
  \item{endDate}{A character that includes the end date
of interest.  Value \emph{must} be in the 
form \code{"yyyy-mm-dd"}, e.g., "2014-01-01".
The function will default to the latest
date in the data frame.                                      }
  \item{\dots}{Arguments to be passed to other functions.  Specifically, this can take
the form of identifying which \code{date} value is of interest.  For example
the sample datasets provided all had either at least \code{assigned_at} 
or \code{created_at} values.  The \code{\link{AssignDateAndDay}} can used
either of these values as the date of interest. }
}
\details{This functions converts values that are \code{nan.0} to \code{NA}.

Boxes in the plot that is produced from this function can help
managers understand the number of reopens by date.  If the 
notches in a box overlap the heavy bar (i.e., the median resolution time)
in an adjacent box, then the resolution time for that priority level is \emph{not}
significantly different than the resolution time in the adjacent plot.

The red dots on a plot indicate outliers in the data set.  In the 
\code{\link{geom_boxplot}} documentation, outliers are defined as data points 
fall outside of \emph{1.5 * IQR} where \emph{IQR} stands for the "Inter-Quartile
Range" of the data.

Plots are generated by the function \code{\link{ggplot}}}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





