\name{PlotReopensByPriority}
\alias{PlotReopensByPriority}
\title{Plot hours to resolution time by priority }
\description{This function plots the number of reopens and produces a box 
plot of those values as a function of \code{priority} level. }
\usage{PlotReopensByPriority(dF)}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by priority level.}
}
\details{This functions converts values that are \code{nan.0} to \code{NA}.

Priority levels in some of the sample data are represented by five 
different values: \code{High}, \code{Low}, \code{Normal}, \code{Urgent}, 
and \code{-}. Review of the sample data showed that the \code{-} priority 
level could not be determined to match any other level.

Boxes in the plot that is produced from this function can help
managers understand the number of reopens by \code{priority} level.  If the 
notches in a box overlap the heavy bar (i.e., the median number of reopens)
in an adjacent box, then the number of reopens for that priority level is \emph{not}
significantly different than the number of reopens in the adjacent box.  This 
provides a quick, easy method of determining if \code{priority} level really 
has an affect on the number of reopens.

The red dots on a plot indicate outliers in the data set.  In the 
\code{\link{geom_boxplot}} documentation, outliers are defined as data points 
fall outside of \emph{1.5 * IQR} where \emph{IQR} stands for the "Inter-Quartile
Range" of the data.

Plots are generated by the function \code{\link{ggplot}}}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





