\name{PlotResolutionTimeByAgent}
\alias{PlotResolutionTimeByAgent}
\title{Create a plot of the resolution time by date}
\description{This function plots the resolution time (in hours) and produces a box 
plot of those values for each \code{assignee_name} in the \code{dF}. The 
call volume of an agent is produced. }
\usage{PlotResolutionTimeByAgent(dF, resTime = "first", ...)}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by date level.}
  \item{resTime}{A character vector that indicates which resolution time
within business hours to plot.  Other values include 
\code{"full"}.}
  \item{\dots}{Arguments to be passed to other functions.  Specifically, this can take
the form of identifying which \code{date} value is of interest.  For example
the sample datasets provided all had either at least \code{assigned_at} 
or \code{created_at} values.  The \code{\link{AssignDateAndDay}} can used
either of these values as the date of interest.}
}
\details{Resolution time in all sample dataFrames is \code{factor}; this function
converts apparently numeric values to \code{as.numeric}.

This functions converts values that are \code{nan.0} to \code{NA}.

The red dots on a plot indicate outliers in the data set.  In the 
\code{\link{geom_boxplot}} documentation, outliers are defined as data points 
fall outside of \emph{1.5 * IQR} where \emph{IQR} stands for the "Inter-Quartile
Range" of the data.

Plots are generated by the function \code{\link{ggplot}}

The blue line in the top plot of the window is the number calls 
received by each \code{assignee_name}.  Assignees with high volume and low 
resolution times are the most efficient agents.  The combination of these 
plots can tell a manager which employees are most efficient at resolving 
calls.

The red line in the bottom plot of the window produced from this function 
is the mean resolution-time value for the entire data frame.  Agents whose 
medians are below this line may be more effective at resolving customer 
issues.  Further, agents that are more effective at resolving issues could 
be considered as 'trainers' for new staff.}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





