\name{PlotResolutionTimeByReopens}
\alias{PlotResolutionTimeByReopens}
\title{Plot hours to resolution time by the number of reopens }
\description{This function plots the chosen value of resolution time (in 
hours) and produces a plot of those values as a function of \code{reopens}. }
\usage{PlotResolutionTimeByReopens(dF, resTime = "first")}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by priority level.}
  \item{resTime}{A character vector that indicates which resolution time
to plot.  Other values include \code{"full"}.}
}
\details{Resolution time in all sample dataFrames is \code{factor}; this function
converts apparently numeric values to \code{as.numeric}

This functions converts values that are \code{nan.0} to \code{NA}.

Boxes in the plot that is produced from this function can help
managers understand the resolution time of \code{reopens} level.  If the 
notches in a box overlap the heavy bar (i.e., the median resolution time)
in an adjacent box, then the resolution time for that number of reopens is \emph{not}
significantly different than the resolution time in the adjacent plot.  This 
provides a quick, easy method of determining if \code{priority} level really 
has an affect on resolution time.

The red dots on a plot indicate outliers in the data set.  In the 
\code{\link{geom_boxplot}} documentation, outliers are defined as data points 
fall outside of \emph{1.5 * IQR} where \emph{IQR} stands for the "Inter-Quartile
Range" of the data.

Plots are generated by the function \code{\link{ggplot}}}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





