\name{PlotScoreByCallsPerPerson}
\alias{PlotScoreByCallsPerPerson}
\title{Create a plot of the score by calls/person}
\description{This function plots the satisfaction score and produces a plot 
of those values as a function of calls/person.  Satisfaction scores are averaged
over the course of a day and the calls/person are calculated over that same 
day.}
\usage{PlotScoreByCallsPerPerson(dF, ...)}
\arguments{
  \item{dF}{The dataFrame for which to convert \code{satisfaction_score} into a
number.}
  \item{\dots}{Arguments to be passed to other functions.  Specifically, this can take
the form of identifying which \code{date} value is of interest.  For example
the sample datasets provided all had either at least \code{assigned_at} 
or \code{created_at} values.  The \code{\link{AssignDateAndDay}} can used
either of these values as the date of interest. }
}
\details{This function calls the \code{\link{ConvertSatisfactionScoreToNumber}} 
function to assign the \code{day} values used.  

Plots are generated by the function \code{\link{ggplot}}

This function uses a linear model to evaluate the relationship between
satisfaction score and the number of calls per person.  If the red line 
(the result of the \code{\link{lm}} in the produced plot trends down, then 
the number of calls per person has a negative affect on satisfaction score.  
If the red line is parallel or very near to parallel, then there is no 
relationship between satisfaction score and the number of calls per person.}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





