\name{PlotScoreByResolutionTime}
\alias{PlotScoreByResolutionTime}
\title{Plot the numeric satisfaction score as a function of wait time}
\description{This function plots the numeric value of satisfaction score 
(calculated by \code{\link{ConvertSatisfactionScoreToNumber}} as a function 
of resolution time.}
\usage{PlotScoreByResolutionTime(dF, resTime = "first")}
\arguments{
  \item{dF}{The dataFrame for which to convert \code{satisfaction_score} into a
number.}
  \item{resTime}{A character vector that indicates which resolution time
to plot.  Other values include \code{"full"}.}
}
\details{This function calls the \code{\link{ConvertSatisfactionScoreToNumber}} 
function to assign the \code{day} values used.  

Plots are generated by the function \code{\link{ggplot}}

This function uses a linear model to evaluate the relationship between
satisfaction score and resolution time.  If the red line (the result of the 
\code{\link{lm}} in the produced plot trends down, then resolution time has 
a negative affect on satisfaction score.  If the red line is parallel or 
very near to parallel, then there is no relationship between satisfaction 
score and resolution time.

In most cases, there are very few or no satisfaction scores.  As a 
result, these data are sparse.}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





