\name{PlotScoreByWaitTime}
\alias{PlotScoreByWaitTime}
\title{Plot the numeric satisfaction score as a function of wait time}
\description{This function plots the numeric value of satisfaction score 
(calculated by \code{\link{ConvertSatisfactionScoreToNumber}} as a function 
of wait time.}
\usage{PlotScoreByWaitTime(dF)}
\arguments{
  \item{dF}{The dataFrame for which to convert \code{satisfaction_score} into a
number.}
}
\details{This function calls the \code{\link{ConvertSatisfactionScoreToNumber}} 
function to assign the \code{day} values used.  

Plots are generated by the function \code{\link{ggplot}}

This function uses a linear model to evaluate the relationship between
satisfaction score and wait time.  If the red line (the result of the \code{\link{lm}}
in the produced plot trends down, then wait time has a negative affect on 
satisfaction score.  If the red line is parallel or very near to parallel, 
then there is relationship between satisfaction score and wait time.}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





