\name{PlotWaitTimeByDay}
\alias{PlotWaitTimeByDay}
\title{Plot hours waited by day of week }
\description{This function plots the wait time (in hours) and produces a box 
plot of those values as a function of \code{day}.  Day levels include those 
assigned from the \code{\link{AssignDateAndDay}} function. }
\usage{PlotWaitTimeByDay(dF, ...)}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by priority level.}
  \item{\dots}{Arguments to be passed to other functions.  Specifically, this can take
the form of identifying which \code{date} value is of interest.  For example
the sample datasets provided all had either at least \code{assigned_at} 
or \code{created_at} values.  The \code{\link{AssignDateAndDay}} can used
either of these values as the date of interest.}
}
\details{This function calls the \code{\link{AssignDateAndDay}} function to 
assign the \code{day} values used.  

Wait time in all sample dataFrames is a \code{factor}; this function
converts apparently numeric values to \code{as.numeric}

This functions converts values that are \code{nan.0} to \code{NA} 
values.  Other \code{NA} values will have to be addressed separately.

Day of week levels in some of the sample data are represented by the 
days of the week.  For the purposes of this function, weeks begin on 
\code{Monday} and continue through \code{Sunday}.

Boxes in the plot that is produced from this function can help
managers understand the wait time on any given \code{dayOfWeek}.  If the 
notches in a box overlap the heavy bar (i.e., the median resolution time)
in an adjacent box, then the resolution time for that priority level is \emph{not}
significantly different than the resolution time in the adjacent plot.  This 
provides a quick, easy method of determining if \code{dayOfWeek} really 
has an affect on resolution time.

The red dots on a plot indicate outliers in the data set.  In the 
\code{\link{geom_boxplot}} documentation, outliers are defined as data points 
fall outside of \emph{1.5 * IQR} where \emph{IQR} stands for the "Inter-Quartile
Range" of the data.

Plots are generated by the function \code{\link{ggplot}}}


\author{Steven H. Ranney Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





