\name{PlotResolutionTimeByReopens}
\alias{PlotResolutionTimeByReopens}
\title{Plot hours to resolution time by the number of reopens }
\description{This function plots the chosen value of resolution time (in 
hours) and produces a plot of those values as a function of \code{reopens}. }
\usage{PlotResolutionTimeByReopens(dF, resTime = "first")}
\arguments{
  \item{dF}{The dataFrame for which to plot resolution time by priority level.}
  \item{resTime}{A character vector that indicates which resolution time
to plot.  Other values include \code{"full"}.}
}
\details{Resolution time in all sample dataFrames is \code{factor}; this function
converts apparently numeric values to \code{as.numeric}

This functions converts values that are \code{nan.0} to \code{NA}.

This function uses a linear model to evaluate the relationship between
the resolution time and the number of reopens.  If the red line 
(the result of the \code{\link{lm}} in the produced plot trends down, then 
the higher the call volume the more times a ticket gets reopened.  If the red 
line is parallel or very near to parallel, then there is no relationship 
between the number of reopens and call volume.  In some cases, there may be 
outliers that exert a high amount of leverage on the model that could cause 
suprious results.  Before drawing conclusions, be sure that you understand the 
data.

Plots are generated by the function \code{\link{ggplot}}}


\author{Steven H. Ranney
Contact: \email{Steven.Ranney@gmail.com}
Steven Ranney}





